Changed the data type to be easy to test

+++ src/libKBNinja/seporate_str.h
@@ -2,4 +2,4 @@
 
 #include <string>
 
-void seporate_str(std::string str, int& purpose, double& time, int& errors);
+void seporate_str(std::string str, int& purpose, int& time, int& errors);
